* develop: (66 commits)
  Improve Setting Create/Update window
  Rework User form
  Refactor some of the file processors
  [#5859] Fix issues with upload via manager for FileSystem media source
  [#5856] Implement ability for derivative Resource types to have their own translatable name
  Tweaks to User Group editing page
  [#5860] Fix FC SQL error when user is in no groups
  Fix FC SQL error when user is not in any user groups
  Allow Form Customization to target specific columns in new Resource layout when moving TVs into tabs
  Convert TVs to new resource field format
  Fix url when drag/drop from a s3 media source
  Fix media source creation processor
  Tweak Quick Update/Create Resource to match new Resource panel, add in class_key to qu
  Fix sorting and filtering of online users dashboard widget
  Fix min compression to not compress unnecessary files, max of 5 per iteration
  Fix issue where tv tab was shown if no TVs
  Refactor the Resource editing page to be more clean/usable, use columns, and more condensed while still decreasing complexity
  Convert and properly sort usergroup getnodes processor
  Clean up User Group, Access Policy and Access Policy Template forms
  Cleanup the dashboards UI
  ...

Conflicts:
	core/docs/changelog.txt
	manager/assets/modext/widgets/element/modx.panel.chunk.js
	manager/assets/modext/widgets/element/modx.panel.plugin.js
	manager/assets/modext/widgets/element/modx.panel.snippet.js
	manager/assets/modext/widgets/element/modx.panel.template.js
	manager/assets/modext/widgets/element/modx.panel.tv.js

+++ core/model/modx/sources/modmediasource.class.php
@@ -530,9 +530,6 @@ class modMediaSource extends modAccessibleSimpleObject {
                     ':context' => $context,
                 );
                 $query = new xPDOCriteria($this->xpdo, $sql, $bindings);
-                if ($this->get('id') == 8) {
-                    $query->prepare($bindings);
-                }
                 if ($query->stmt && $query->stmt->execute()) {
                     while ($row = $query->stmt->fetch(PDO::FETCH_ASSOC)) {
                         $policy['sources.modAccessMediaSource'][$row['target']][] = array(
@@ -591,24 +588,20 @@ class modMediaSource extends modAccessibleSimpleObject {
         if (empty($cacheManager)) return;
 
         $c = $this->xpdo->newQuery('modContext');
-        $c->select($this->xpdo->getSelectColumns('modContext','modContext','',array('key')));
-        $c->prepare();
-        $sql = $c->toSQL();
+        $c->select($this->xpdo->escape('key'));
 
-        $options[xPDO::OPT_CACHE_KEY] = $this->getOption('cache_context_settings_key', $options, 'context_settings');
+        $options[xPDO::OPT_CACHE_KEY] = $this->getOption('cache_media_sources_key', $options, 'media_sources');
         $options[xPDO::OPT_CACHE_HANDLER] = $this->getOption('cache_media_sources_handler', $options, $this->getOption(xPDO::OPT_CACHE_HANDLER, $options));
         $options[xPDO::OPT_CACHE_FORMAT] = (integer) $this->getOption('cache_media_sources_format', $options, $this->getOption(xPDO::OPT_CACHE_FORMAT, $options, xPDOCacheManager::CACHE_PHP));
         $options[xPDO::OPT_CACHE_ATTEMPTS] = (integer) $this->getOption('cache_media_sources_attempts', $options, $this->getOption(xPDO::OPT_CACHE_ATTEMPTS, $options, 10));
         $options[xPDO::OPT_CACHE_ATTEMPT_DELAY] = (integer) $this->getOption('cache_media_sources_attempt_delay', $options, $this->getOption(xPDO::OPT_CACHE_ATTEMPT_DELAY, $options, 1000));
 
-        $stmt = $this->xpdo->query($sql);
-        if ($stmt) {
-            while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+        if ($c->prepare() && $c->stmt->execute()) {
+            while ($row = $c->stmt->fetch(PDO::FETCH_ASSOC)) {
                 if ($row && !empty($row['key'])) {
                     $cacheManager->delete($row['key'].'/source',$options);
                 }
             }
-            $stmt->closeCursor();
         }
     }
 }
